fix: satisfy static format and max-lines checks
Format the search panel CustomScrollView tree, move simple panel wiring out of the shell body, and extract the source control scroll layout so those files stay within the line ratchet.
