+++ lib/src/features/shell/presentation/alera_shell_page_body.dart
@@ -144,179 +144,17 @@ class _AleraShellPageBodyState extends ConsumerState<_AleraShellPageBody> {
                                           ? maximumPanelWidth
                                           : AleraTokens.sidebarMaxWidth,
                                       panelBuilder: simple
-                                          ? (
-                                              toolFor,
-                                            ) => SimpleWorkspacePanelView(
-                                              workspaceId: workspace.id,
+                                          ? (toolFor) => _buildSimplePanelView(
+                                              workspace: workspace,
+                                              project: project,
                                               panel: panel,
+                                              sourceControlScope:
+                                                  sourceControlScope,
                                               tabs: shell.tabs,
-                                              tabBuilder:
-                                                  (tab, active, groupId) =>
-                                                      _buildSimplePanelTab(
-                                                        workspace: workspace,
-                                                        panel: panel,
-                                                        tabs: shell.tabs,
-                                                        tab: tab,
-                                                        active: active,
-                                                        groupId: groupId,
-                                                      ),
-                                              onSelect: (key) => controller
-                                                  .selectSimplePanelKey(
-                                                    workspace.id,
-                                                    key,
-                                                  ),
-                                              onSelectInGroup: (groupId, key) =>
-                                                  controller
-                                                      .selectSimplePanelKey(
-                                                        workspace.id,
-                                                        key,
-                                                        groupId: groupId,
-                                                      ),
-                                              onClose: (key) async {
-                                                final tool =
-                                                    SimpleWorkspaceTool.forKey(
-                                                      key,
-                                                    );
-                                                if (tool != null) {
-                                                  controller.closeSimpleTool(
-                                                    workspace.id,
-                                                    tool,
-                                                  );
-                                                } else if (SimpleWorkspacePanel.tabId(
-                                                      key,
-                                                    )
-                                                    case final String id) {
-                                                  if (await _confirmCloseDirtyTabs(
-                                                    shell.tabs,
-                                                    <String>[id],
-                                                  )) {
-                                                    await controller
-                                                        .closeWorkspaceTab(
-                                                          workspace: workspace,
-                                                          tabId: id,
-                                                        );
-                                                  }
-                                                }
-                                              },
-                                              onNewTerminal: () => unawaited(
-                                                controller.createTerminalTab(
-                                                  workspace,
-                                                ),
-                                              ),
-                                              onNewTerminalInGroup: (groupId) {
-                                                unawaited(
-                                                  controller.createTerminalTab(
-                                                    workspace,
-                                                    targetGroupId: groupId,
-                                                  ),
-                                                );
-                                              },
-                                              onSplitGroup: (groupId, zone) {
-                                                unawaited(
-                                                  controller
-                                                      .splitWorkbenchGroupWithTerminal(
-                                                        workspace: workspace,
-                                                        groupId: groupId,
-                                                        zone: zone,
-                                                      ),
-                                                );
-                                              },
-                                              onMergeGroup: (groupId) {
-                                                unawaited(
-                                                  controller
-                                                      .mergeWorkbenchGroupIntoSibling(
-                                                        workspaceId:
-                                                            workspace.id,
-                                                        groupId: groupId,
-                                                      ),
-                                                );
-                                              },
-                                              onMoveTab:
-                                                  ({
-                                                    required key,
-                                                    required targetGroupId,
-                                                    required zone,
-                                                    index,
-                                                  }) {
-                                                    unawaited(
-                                                      controller
-                                                          .moveWorkspaceTab(
-                                                            workspaceId:
-                                                                workspace.id,
-                                                            tabId: key,
-                                                            targetGroupId:
-                                                                targetGroupId,
-                                                            zone: zone,
-                                                            index: index,
-                                                          ),
-                                                    );
-                                                  },
-                                              onUpdateSplitRatio: (path, ratio) {
-                                                controller
-                                                    .updateWorkbenchSplitRatio(
-                                                      workspaceId: workspace.id,
-                                                      nodePath: path,
-                                                      ratio: ratio,
-                                                    );
-                                              },
-                                              onHide: controller
-                                                  .toggleRightSidebarVisible,
-                                              surfaceBuilder: (key) {
-                                                final tool =
-                                                    SimpleWorkspaceTool.forKey(
-                                                      key,
-                                                    );
-                                                if (tool != null) {
-                                                  return Focus(
-                                                    canRequestFocus: false,
-                                                    onFocusChange: (focused) {
-                                                      if (focused) {
-                                                        controller
-                                                            .selectSimplePanelKey(
-                                                              workspace.id,
-                                                              key,
-                                                            );
-                                                      }
-                                                    },
-                                                    child: toolFor(switch (tool) {
-                                                      SimpleWorkspaceTool
-                                                          .explorer =>
-                                                        WorkbenchContextPanelTab
-                                                            .explorer,
-                                                      SimpleWorkspaceTool
-                                                          .search =>
-                                                        WorkbenchContextPanelTab
-                                                            .search,
-                                                      SimpleWorkspaceTool
-                                                          .sourceControl =>
-                                                        WorkbenchContextPanelTab
-                                                            .gitDiff,
-                                                      SimpleWorkspaceTool
-                                                          .pullRequest =>
-                                                        WorkbenchContextPanelTab
-                                                            .pullRequests,
-                                                    }),
-                                                  );
-                                                }
-                                                return _buildContent(
-                                                  bootstrapped:
-                                                      shell.bootstrapped,
-                                                  hasProjects:
-                                                      shell.hasProjects,
-                                                  project: project,
-                                                  workspace: workspace,
-                                                  sourceControlScope:
-                                                      sourceControlScope,
-                                                  tabs: shell.tabs,
-                                                  layout: shell.layout,
-                                                  singleSurface: true,
-                                                  singleTabId:
-                                                      SimpleWorkspacePanel.tabId(
-                                                        key,
-                                                      ),
-                                                );
-                                              },
-                                              content: const SizedBox.shrink(),
+                                              bootstrapped: shell.bootstrapped,
+                                              hasProjects: shell.hasProjects,
+                                              layout: shell.layout,
+                                              toolFor: toolFor,
                                             )
                                           : null,
                                       sourceControlScope: sourceControlScope,
